refactor(lambda): split service.rs core handlers by surface (2866 -> 1430 mod)
Audit 2026-05-19 P9 continuation. Lambda's service.rs core split:

- mod.rs        1430  (AwsService trait dispatch + iam_* helpers + struct
                       defs/builders + LambdaInvocationCleanup drop)
- functions.rs        (create / get / delete / list function)
- invoke.rs           (invoke + lookup_destination_config)
- publish.rs          (publish_version + function_config_json)
- init.rs             (resolve_action helper)

151 unit tests pass.
